# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Erdos 服务端 API 契约
  version: 0.1.0-draft
  description: >
    客户端 / 引擎 / 服务端三方的共同语言。契约冻结项：错误码（x-error-codes）、
    表结构、签名格式——冻结后变更必须走契约评审（研发手册「质量标准」红线）。
    本文件为草案（x-contract-status: draft）：已 100% 覆盖实际实现（56 URL，2026-10-02
    契约一致性自检），待 ARCH 评审冻结为 v1。自检入口：server/scripts/contract_audit.py。
  x-contract-status: draft
servers:
  - url: https://api.erdos.example.com
...
paths:
...
  /v1/points/balance:
    get:
      tags: [points]
      operationId: getBalance
      summary: 积分账户余额
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 双余额（购买/月度）+ 冻结标记
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvelopeBalanceView" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/points/reserve:
    post:
      tags: [points]
      operationId: reservePoints
      summary: 预扣（阶段许可 Ed25519 签发）
      description: exec_id 幂等；余额不足 40901；欠费冻结 40302；客户端凭许可离线验签启动阶段。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReserveBody" }
      responses:
        "200":
          description: 扣减后余额 + 阶段许可
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvelopeReserveView" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: 账号欠费冻结（ACCOUNT_FROZEN）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
        "409":
          description: 余额不足或状态冲突（CONFLICT）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
  /v1/points/confirm:
    post:
      tags: [points]
      operationId: confirmPoints
      summary: 确认消耗（阶段完成，终态）
      description: reserved → confirmed；重复确认幂等返回既有终态；退还后确认 40901。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SettleBody" }
      responses:
        "200":
          description: 余额 + 终态状态
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvelopeSettleView" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: 终态迁移冲突（CONFLICT）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
  /v1/points/refund:
    post:
      tags: [points]
      operationId: refundPoints
      summary: 退还（阶段取消，返还余额）
      description: reserved → refunded（终态），原子返还余额；重复退还幂等。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SettleBody" }
      responses:
        "200":
          description: 余额 + 终态状态
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvelopeSettleView" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: 终态迁移冲突（CONFLICT）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
  /v1/points/offline-sync:
    post:
      tags: [points]
      operationId: offlineSyncPoints
      summary: 离线消耗批量对账（逐条回执）
      description: 幂等去重（exec_id）；余额不足逐条 insufficient；批次后总余额 ≤ 阈值才冻结。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OfflineSyncBody" }
      responses:
        "200":
          description: 应用/重复/不足计数 + 冻结状态 + 逐条回执
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvelopeOfflineSyncView" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/points/ledger:
    get:
      tags: [points]
      operationId: getLedger
      summary: 积分流水账单（分页）
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: 流水列表 + 总数（按时间倒序）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvelopeLedgerPageView" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/points/ledger/export:
    get:
      tags: [points]
      operationId: exportLedgerCsv
      summary: 积分流水 CSV 导出
      description: UTF-8 BOM CSV（StreamingResponse），上限与分页一致。
      security:
        - BearerAuth: []
      responses:
        "200":
          description: CSV 文件流
          content:
            text/csv:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT 双令牌（SP2-2 接入；SP2-1 网关仅做鉴权透传）
  responses:
    Unauthorized:
      description: 未认证（UNAUTHENTICATED）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Envelope" }
...
    NotFound:
      description: 资源不存在（NOT_FOUND）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Envelope" }
    BadRequest:
      description: 参数错误（BAD_REQUEST）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Envelope" }
...
  schemas:
...
    BalanceView:
      type: object
      required: [user_id, purchased_balance, monthly_balance, frozen]
      properties:
        user_id: { type: string }
        purchased_balance: { type: integer }
        monthly_balance: { type: integer }
        frozen: { type: boolean }
    LedgerView:
      type: object
      required: [id, exec_id, delta, balance_type, kind, status, source, created_at]
      properties:
        id: { type: string }
        exec_id: { type: string }
        delta: { type: integer, description: 入账为正，扣减为负 }
        balance_type: { type: string, enum: [purchased, monthly] }
        kind: { type: string, enum: [grant, reserve, confirm, refund, offline_sync] }
        status: { type: string, enum: [reserved, confirmed, refunded] }
        source: { type: string }
        stage: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    LedgerPageView:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/LedgerView" }
        total: { type: integer }
    ReserveBody:
      type: object
      required: [exec_id, stage, points]
      properties:
        exec_id: { type: string, minLength: 1, maxLength: 64 }
        task_id: { type: string, maxLength: 36, nullable: true }
        stage: { type: string, minLength: 1, maxLength: 16 }
        points: { type: integer, exclusiveMinimum: 0 }
    GrantView:
      type: object
      required: [exec_id, stage, points, signature, key_version, issued_at, expires_at, status]
      properties:
        exec_id: { type: string }
        stage: { type: string }
        points: { type: integer }
        signature: { type: string }
        key_version: { type: string }
        issued_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        status: { type: string, enum: [active, used, expired] }
    ReserveView:
      type: object
      required: [balance, grant, already_reserved]
      properties:
        balance: { $ref: "#/components/schemas/BalanceView" }
        grant: { $ref: "#/components/schemas/GrantView" }
        already_reserved: { type: boolean }
    SettleBody:
      type: object
      required: [exec_id]
      properties:
        exec_id: { type: string, minLength: 1, maxLength: 64 }
    SettleView:
      type: object
      required: [balance, status]
      properties:
        balance: { $ref: "#/components/schemas/BalanceView" }
        status: { type: string, enum: [confirmed, refunded] }
    OfflineSyncBody:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 200
          items:
            type: object
            required: [exec_id, stage, points]
            properties:
              exec_id: { type: string, minLength: 1, maxLength: 64 }
              task_id: { type: string, maxLength: 36, nullable: true }
              stage: { type: string, minLength: 1, maxLength: 16 }
              points: { type: integer, exclusiveMinimum: 0 }
    OfflineItemView:
      type: object
      required: [exec_id, status, detail]
      properties:
        exec_id: { type: string }
        status: { type: string, enum: [applied, duplicate, insufficient] }
        detail: { type: string }
    OfflineSyncView:
      type: object
      required: [applied, duplicate, insufficient, frozen, items]
      properties:
        applied: { type: integer }
        duplicate: { type: integer }
        insufficient: { type: integer }
        frozen: { type: boolean }
        balance: { $ref: "#/components/schemas/BalanceView" }
        items:
          type: array
          items: { $ref: "#/components/schemas/OfflineItemView" }
    EnvelopeBalanceView:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/BalanceView" }
    EnvelopeLedgerPageView:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/LedgerPageView" }
    EnvelopeReserveView:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/ReserveView" }
    EnvelopeSettleView:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/SettleView" }
    EnvelopeOfflineSyncView:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/OfflineSyncView" }

    # ---- 权益与内容域（SP2-6）----
...
    Envelope:
      type: object
      description: 统一响应信封。code=0 表示成功；其它取值与 x-error-codes 一一对应。
      required: [code, message, request_id, timestamp]
      properties:
        code:
          type: integer
          description: 业务码，0 表示成功，非 0 见 x-error-codes
        message:
          type: string
          description: 面向使用者的提示文案
        detail:
          type: string
          description: 可选的错误上下文（仅失败时携带）
          nullable: true
        data:
          description: 业务数据（成功时携带）
          nullable: true
        request_id:
          type: string
          description: 请求追踪 ID，与响应头 X-Request-Id 一致
        timestamp:
          type: string
          format: date-time
          description: 服务端 UTC 时间
...